requirements.txt
dateparser==1.2.2
discord.py==2.6.4
faiss_cpu==1.11.0
Flask==3.1.2
google_api_python_client==2.187.0
google_auth_oauthlib==1.2.3
numpy==2.4.0
openai==2.14.0
pandas==2.3.3
python-dotenv==1.2.1
sentence_transformers==5.2.0
torch==2.9.1
